rag/requirements.txt
# 核心框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
python-multipart==0.0.6

# LangChain相关
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.3

# 向量数据库
pymilvus==2.3.5

# 缓存
redis==5.0.1

# 文本处理
jieba==0.42.1
rank-bm25==0.2.2

# LLM相关
openai==1.10.0

# 工具和辅助
httpx==0.26.0
python-dotenv==1.0.0
